Further refactoring after local testing
