+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/esa/EsaTemplateContent.java
@@ -127,6 +127,7 @@ public abstract class EsaTemplateContent extends WriteFinalDecisionTemplateConte
             }
         }
     }
+
     public void addAnythingElseIfPresent(WriteFinalDecisionTemplateBody writeFinalDecisionTemplateBody) {
         if (writeFinalDecisionTemplateBody.getAnythingElse() != null) {
             addComponent(new Paragraph(EsaTemplateComponentId.ANYTHING_ELSE.name(), writeFinalDecisionTemplateBody.getAnythingElse()));
