+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/esa/EsaWriteFinalDecisionPreviewDecisionService.java
@@ -1,0 +1,133 @@
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.esa;
+
+import static java.util.Optional.empty;
+import static org.apache.commons.lang3.StringUtils.join;
+import static org.apache.commons.lang3.StringUtils.splitByCharacterTypeCamelCase;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Optional;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.idam.client.IdamClient;
+import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
+import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.AwardType;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.WriteFinalDecisionPreviewDecisionServiceBase;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.esa.scenarios.EsaScenario;
+import uk.gov.hmcts.reform.sscs.config.DocumentConfiguration;
+import uk.gov.hmcts.reform.sscs.docassembly.GenerateFile;
+import uk.gov.hmcts.reform.sscs.model.docassembly.Descriptor;
+import uk.gov.hmcts.reform.sscs.model.docassembly.NoticeIssuedTemplateBody.NoticeIssuedTemplateBodyBuilder;
+import uk.gov.hmcts.reform.sscs.model.docassembly.WriteFinalDecisionTemplateBody;
+import uk.gov.hmcts.reform.sscs.model.docassembly.WriteFinalDecisionTemplateBody.WriteFinalDecisionTemplateBodyBuilder;
+import uk.gov.hmcts.reform.sscs.service.DecisionNoticeOutcomeService;
+import uk.gov.hmcts.reform.sscs.service.EsaDecisionNoticeOutcomeService;
+import uk.gov.hmcts.reform.sscs.service.EsaDecisionNoticeQuestionService;
+
+@Slf4j
+@Component
+public class EsaWriteFinalDecisionPreviewDecisionService extends WriteFinalDecisionPreviewDecisionServiceBase {
+
+    private EsaDecisionNoticeQuestionService esaDecisionNoticeQuestionService;
+
+    @Autowired
+    public EsaWriteFinalDecisionPreviewDecisionService(GenerateFile generateFile, IdamClient idamClient,
+        EsaDecisionNoticeQuestionService decisionNoticeQuestionService, EsaDecisionNoticeOutcomeService outcomeService, DocumentConfiguration documentConfiguration) {
+        super(generateFile, idamClient, decisionNoticeQuestionService, outcomeService, documentConfiguration);
+        this.esaDecisionNoticeQuestionService = decisionNoticeQuestionService;
+    }
+
+    @Override
+    public String getBenefitType() {
+        return "ESA";
+    }
+
+    @Override
+    protected void setTemplateContent(DecisionNoticeOutcomeService outcomeService, PreSubmitCallbackResponse<SscsCaseData> response,
+        NoticeIssuedTemplateBodyBuilder builder, SscsCaseData caseData,
+        WriteFinalDecisionTemplateBody payload) {
+
+        // Validate here for ESA instead of only validating on submit.
+        // This ensures that we know we can obtain a valid allowed or refused condition below.
+        outcomeService.validate(response, caseData);
+
+        if (response.getErrors().isEmpty()) {
+
+            // If validation has produced no errors, we know that we can get an allowed/refused condition.
+            Optional<EsaAllowedOrRefusedCondition> condition = EsaPointsRegulationsAndSchedule3ActivitiesCondition
+                .getPassingAllowedOrRefusedCondition(decisionNoticeQuestionService, caseData);
+            if (condition.isPresent()) {
+                EsaScenario scenario = condition.get().getEsaScenario(caseData);
+                EsaTemplateContent templateContent = scenario.getContent(payload);
+                builder.writeFinalDecisionTemplateContent(templateContent);
+            } else {
+                // Should never happen.
+                response.addError("Unable to obtain a valid scenario - something has gone wrong");
+            }
+        } else {
+            Iterator<String> iter = response.getErrors().iterator();
+            while (iter.hasNext()) {
+                System.out.println(iter.next());
+            }
+
+        }
+    }
+
+    @Override
+    protected void setEntitlements(WriteFinalDecisionTemplateBodyBuilder builder, SscsCaseData caseData) {
+
+        builder.esaIsEntited(false);
+        builder.esaAwardRate(null);
+        Optional<AwardType> esaAwardTypeOptional = caseData.isWcaAppeal() ? EsaPointsRegulationsAndSchedule3ActivitiesCondition
+            .getTheSinglePassingPointsConditionForSubmittedActivitiesAndPoints(decisionNoticeQuestionService, caseData).getAwardType() : empty();
+        if (!esaAwardTypeOptional.isEmpty()) {
+            String esaAwardType = esaAwardTypeOptional.get().getKey();
+            if (esaAwardType != null) {
+                builder.esaAwardRate(join(
+                    splitByCharacterTypeCamelCase(esaAwardType), ' ').toLowerCase());
+            }
+
+            if (AwardType.LOWER_RATE.getKey().equals(esaAwardType)
+                || AwardType.HIGHER_RATE.getKey().equals(esaAwardType)) {
+                builder.esaIsEntited(true);
+            }
+        }
+    }
+
+    protected List<Descriptor> getEsaDescriptorsFromQuestionKeys(SscsCaseData caseData, List<String> questionKeys) {
+        return getDescriptorsFromQuestionKeys("ESA", key -> esaDecisionNoticeQuestionService.extractQuestionFromKey(EsaActivityQuestionKey.getByKey(key)), caseData, questionKeys);
+    }
+
+    @Override
+    protected void setDescriptorsAndPoints(WriteFinalDecisionTemplateBodyBuilder builder, SscsCaseData caseData) {
+        List<Descriptor> allDescriptors = new ArrayList<>();
+        List<String> physicalDisabilityAnswers = EsaActivityType.PHYSICAL_DISABILITIES.getAnswersExtractor().apply(caseData);
+        if (physicalDisabilityAnswers != null) {
+            List<Descriptor> physicalDisablityDescriptors = getEsaDescriptorsFromQuestionKeys(caseData, physicalDisabilityAnswers);
+            allDescriptors.addAll(physicalDisablityDescriptors);
+        }
+        List<String> mentalAssessmentAnswers = EsaActivityType.MENTAL_ASSESSMENT.getAnswersExtractor().apply(caseData);
+        if (mentalAssessmentAnswers != null) {
+            List<Descriptor> mentalAssessmentDescriptors = getEsaDescriptorsFromQuestionKeys(caseData, mentalAssessmentAnswers);
+            allDescriptors.addAll(mentalAssessmentDescriptors);
+        }
+
+        if (allDescriptors.isEmpty()) {
+            builder.esaSchedule2Descriptors(null);
+            builder.esaNumberOfPoints(null);
+        } else {
+            builder.esaSchedule2Descriptors(allDescriptors);
+            int numberOfPoints = allDescriptors.stream().mapToInt(Descriptor::getActivityAnswerPoints).sum();
+            if (EsaPointsCondition.POINTS_GREATER_OR_EQUAL_TO_FIFTEEN.getPointsRequirementCondition().test(numberOfPoints)) {
+                caseData.setDoesRegulation29Apply(null);
+            }
+            builder.esaNumberOfPoints(numberOfPoints);
+        }
+        builder.regulation29Applicable(caseData.getDoesRegulation29Apply() == null ? null :  caseData.getDoesRegulation29Apply().toBoolean());
+        builder.regulation35Applicable(caseData.getDoesRegulation35Apply() == null ? null :  caseData.getDoesRegulation35Apply().toBoolean());
+        builder.supportGroupOnly(caseData.isSupportGroupOnlyAppeal());
+    }
+}
