+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/pip/PipWriteFinalDecisionMidEventValidationHandler.java
@@ -1,0 +1,118 @@
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.pip;
+
+import static org.apache.commons.lang3.StringUtils.equalsIgnoreCase;
+
+import javax.validation.Validator;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
+import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.domain.YesNo;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.AwardType;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.WriteFinalDecisionMidEventValidationHandlerBase;
+import uk.gov.hmcts.reform.sscs.service.DecisionNoticeService;
+
+@Slf4j
+@Component
+public class PipWriteFinalDecisionMidEventValidationHandler extends WriteFinalDecisionMidEventValidationHandlerBase {
+
+    public PipWriteFinalDecisionMidEventValidationHandler(Validator validator, DecisionNoticeService decisionNoticeService) {
+        super(validator, decisionNoticeService);
+    }
+
+    @Override
+    protected String getBenefitType() {
+        return "PIP";
+    }
+
+    @Override
+    protected void setDefaultFields(SscsCaseData sscsCaseData) {
+        if (sscsCaseData.getWriteFinalDecisionEndDateType() == null && "yes".equalsIgnoreCase(sscsCaseData.getWriteFinalDecisionIsDescriptorFlow())) {
+            if (bothDailyLivingAndMobilityQuestionsAnswered(sscsCaseData) && isNoAwardOrNotConsideredForDailyLiving(sscsCaseData)
+                && isNoAwardOrNotConsideredForMobility(sscsCaseData)) {
+                sscsCaseData.setWriteFinalDecisionEndDateType("na");
+            }
+        }
+    }
+
+    @Override
+    protected void setShowPageFlags(SscsCaseData sscsCaseData) {
+        // N/A for PIP
+    }
+
+    @Override
+    protected void validateAwardTypes(SscsCaseData sscsCaseData, PreSubmitCallbackResponse<SscsCaseData> preSubmitCallbackResponse) {
+        if ((!equalsIgnoreCase(AwardType.NO_AWARD.getKey(), sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion())
+            || !equalsIgnoreCase(AwardType.NO_AWARD.getKey(), sscsCaseData.getPipWriteFinalDecisionMobilityQuestion()))
+            &&  sscsCaseData.getPipWriteFinalDecisionDailyLivingActivitiesQuestion() != null
+            &&  sscsCaseData.getPipWriteFinalDecisionMobilityActivitiesQuestion() != null
+            &&  sscsCaseData.getPipWriteFinalDecisionDailyLivingActivitiesQuestion().isEmpty()
+            &&  sscsCaseData.getPipWriteFinalDecisionMobilityActivitiesQuestion().isEmpty()) {
+            preSubmitCallbackResponse.addError("At least one activity must be selected unless there is no award");
+        }
+
+        if (AwardType.NO_AWARD.getKey().equals(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion())
+            && "higher".equals(sscsCaseData.getPipWriteFinalDecisionComparedToDwpDailyLivingQuestion())) {
+            preSubmitCallbackResponse.addError("Daily living decision of No Award cannot be higher than DWP decision");
+        }
+
+        if (AwardType.NO_AWARD.getKey().equals(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion())
+            && "higher".equals(sscsCaseData.getPipWriteFinalDecisionComparedToDwpMobilityQuestion())) {
+            preSubmitCallbackResponse.addError("Mobility decision of No Award cannot be higher than DWP decision");
+
+        }
+        if (AwardType.ENHANCED_RATE.getKey().equals(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion())
+            && "lower".equals(sscsCaseData.getPipWriteFinalDecisionComparedToDwpDailyLivingQuestion())) {
+            preSubmitCallbackResponse.addError("Daily living award at Enhanced Rate cannot be lower than DWP decision");
+        }
+        if (AwardType.ENHANCED_RATE.getKey().equals(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion())
+            && "lower".equals(sscsCaseData.getPipWriteFinalDecisionComparedToDwpMobilityQuestion())) {
+            preSubmitCallbackResponse.addError("Mobility award at Enhanced Rate cannot be lower than DWP decision");
+        }
+
+        if (AwardType.NOT_CONSIDERED.getKey().equals(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion())
+            && AwardType.NOT_CONSIDERED.getKey().equals(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion())) {
+            preSubmitCallbackResponse.addError("At least one of Mobility or Daily Living must be considered");
+        }
+
+        if ("yes".equalsIgnoreCase(sscsCaseData.getWriteFinalDecisionIsDescriptorFlow()) && bothDailyLivingAndMobilityQuestionsAnswered(sscsCaseData)) {
+            if (isNoAwardOrNotConsideredForDailyLiving(sscsCaseData)
+                && isNoAwardOrNotConsideredForMobility(sscsCaseData)) {
+                if (sscsCaseData.getWriteFinalDecisionEndDateType() != null && !"na".equals(sscsCaseData.getWriteFinalDecisionEndDateType())) {
+                    preSubmitCallbackResponse.addError("End date is not applicable for this decision - please specify 'N/A - No Award'.");
+                }
+            } else {
+                if ("na".equals(sscsCaseData.getWriteFinalDecisionEndDateType())) {
+                    preSubmitCallbackResponse.addError("An end date must be provided or set to Indefinite for this decision.");
+                }
+            }
+        }
+    }
+
+    @Override
+    protected void setShowSummaryOfOutcomePage(SscsCaseData sscsCaseData) {
+        if (sscsCaseData.getWriteFinalDecisionIsDescriptorFlow() != null && sscsCaseData.getWriteFinalDecisionIsDescriptorFlow().equalsIgnoreCase(YesNo.NO.getValue())
+            && sscsCaseData.getWriteFinalDecisionGenerateNotice() != null && sscsCaseData.getWriteFinalDecisionGenerateNotice().equalsIgnoreCase(YesNo.YES.getValue())) {
+            sscsCaseData.setShowFinalDecisionNoticeSummaryOfOutcomePage(YesNo.YES);
+            return;
+        }
+        sscsCaseData.setShowFinalDecisionNoticeSummaryOfOutcomePage(YesNo.NO);
+    }
+
+    private boolean isNoAwardOrNotConsideredForMobility(SscsCaseData sscsCaseData) {
+        return sscsCaseData.getPipWriteFinalDecisionMobilityQuestion() != null
+            && ("noAward".equals(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion())
+            || "notConsidered".equals(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion()));
+    }
+
+    private boolean isNoAwardOrNotConsideredForDailyLiving(SscsCaseData sscsCaseData) {
+        return sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion() != null
+            && ("noAward".equals(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion())
+            || "notConsidered".equals(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion()));
+    }
+
+    private boolean bothDailyLivingAndMobilityQuestionsAnswered(SscsCaseData caseData) {
+        return caseData.getPipWriteFinalDecisionDailyLivingQuestion() != null
+            && caseData.getPipWriteFinalDecisionMobilityQuestion() != null;
+    }
+}
