+++ src/main/java/uk/gov/hmcts/reform/sscs/service/DecisionNoticeOutcomeService.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.sscs.service;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.Outcome.DECISION_IN_FAVOUR_OF_APPELLANT;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.Outcome.DECISION_UPHELD;
 
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
