+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/WriteFinalDecisionAboutToSubmitHandlerTestBase.java
@@ -6,14 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.openMocks;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
-import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.DRAFT_DECISION_NOTICE;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import junitparams.Parameters;
 import org.junit.Before;
 import org.junit.Test;
@@ -26,8 +22,6 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.CaseDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocument;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocumentDetails;
 import uk.gov.hmcts.reform.sscs.service.DecisionNoticeOutcomeService;
 import uk.gov.hmcts.reform.sscs.service.DecisionNoticeQuestionService;
 import uk.gov.hmcts.reform.sscs.service.DecisionNoticeService;
@@ -60,7 +54,7 @@ public abstract class WriteFinalDecisionAboutToSubmitHandlerTestBase<T extends D
     @Before
     public void setUp() throws IOException {
         openMocks(this);
-        decisionNoticeService = new DecisionNoticeService(Arrays.asList(decisionNoticeQuestionService), Arrays.asList(createOutcomeService(decisionNoticeQuestionService)));
+        decisionNoticeService = new DecisionNoticeService(Arrays.asList(decisionNoticeQuestionService), Arrays.asList(createOutcomeService(decisionNoticeQuestionService)), Arrays.asList());
         previewDocumentService = new PreviewDocumentService();
         handler = new WriteFinalDecisionAboutToSubmitHandler(decisionNoticeService, previewDocumentService);
 
@@ -186,21 +180,7 @@ public abstract class WriteFinalDecisionAboutToSubmitHandlerTestBase<T extends D
     }
 
     @Test
-    public void givenDraftFinalDecisionAlreadyExistsOnCase_thenOverwriteExistingDraft() {
-        SscsDocument doc = SscsDocument.builder().value(SscsDocumentDetails.builder().documentFileName("oldDraft.doc").documentType(DRAFT_DECISION_NOTICE.getValue()).build()).build();
-        List<SscsDocument> docs = new ArrayList<>();
-        docs.add(doc);
-        callback.getCaseDetails().getCaseData().setSscsDocument(docs);
-
-        sscsCaseData.setWriteFinalDecisionGenerateNotice("yes");
-
-        // Why do we not need to set valid scenario ?
-
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-
-        assertEquals(1, response.getData().getSscsDocument().size());
-        assertEquals((String.format("Draft Decision Notice generated on %s.pdf", LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-YYYY")))), response.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
+    public abstract void givenDraftFinalDecisionAlreadyExistsOnCase_thenOverwriteExistingDraft();
 
     @Test
     @Parameters({"ABOUT_TO_START", "MID_EVENT", "SUBMITTED"})
